requirements.txt
tensorflow>=2.12.0
flask>=2.3.0
werkzeug>=2.3.0
scipy>=1.16.0
matplotlib>=3.7.0
pillow>=10.0.0
numpy>=1.24.0
reportlab>=3.6.12
